Agrego sweetalert a proyectos en parte privada

+++ src/app/componentes/parte-privada/proyectos/proyecto-delete/proyecto-delete.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ProyectoService } from 'src/app/service/parte-privada/proyecto.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-proyecto-delete',
@@ -40,9 +41,30 @@ export class ProyectoDeleteComponent implements OnInit {
   }
 
   eliminarProyecto(){
-    let idInput = document.getElementById('delete-proyecto-input-id') as HTMLInputElement;
-    let idNumerico = Number(idInput.value);
-    this.proyectoService.deleteProyecto(idNumerico).subscribe(data => data);
+
+    Swal.fire({
+      title: 'Estás seguro?',
+      text: "Por favor, confirma los cambios",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        let idInput = document.getElementById('delete-proyecto-input-id') as HTMLInputElement;
+        let idNumerico = Number(idInput.value);
+        this.proyectoService.deleteProyecto(idNumerico).subscribe(data => data);
+        Swal.fire(
+          'Éxito!',
+          'El elemento fue eliminado correctamente.',
+          'success'
+        )
+        this.proyectoForm.reset('');
+      }
+    })
+
+    
   }
 
   initForm():FormGroup {
